Moved custom validation test from ConfigurationValidation file to CustomValidations file. (alle usages of  cfg.Advanced.Validator)
