+++ src/UnitTests/ConfigurationValidation.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using AutoMapper.Mappers;
 using Shouldly;
 using Xunit;
 
@@ -37,69 +36,6 @@ namespace AutoMapper.UnitTests.ConfigurationValidation
         [Fact]
         public void Should_fail_validation() => new Action(Configuration.AssertConfigurationIsValid).ShouldThrowException<AutoMapperConfigurationException>(ex=>
             ex.MemberMap.DestinationName.ShouldBe("AutoMapper.UnitTests.ConfigurationValidation.ConstructorMappingValidation+Destination.Void .ctor(ComplexType).parameter myComplexMember"));
-    }
-
-    public class When_using_custom_validation
-    {
-        bool _calledForRoot = false;
-        bool _calledForValues = false;
-        bool _calledForInt = false;
-
-        public class Source
-        {
-            public int[] Values { get; set; }
-        }
-
-        public class Dest
-        {
-            public int[] Values { get; set; }
-        }
-
-        [Fact]
-        public void Should_call_the_validator()
-        {
-            var config = new MapperConfiguration(cfg =>
-            {
-                cfg.Advanced.Validator(Validator);
-                cfg.CreateMap<Source, Dest>();
-            });
-
-            config.AssertConfigurationIsValid();
-
-            _calledForRoot.ShouldBeTrue();
-            _calledForValues.ShouldBeTrue();
-            _calledForInt.ShouldBeTrue();
-        }
-
-        private void Validator(ValidationContext context)
-        {
-            if(context.TypeMap != null)
-            {
-                _calledForRoot = true;
-                context.TypeMap.Types.ShouldBe(context.Types);
-                context.Types.SourceType.ShouldBe(typeof(Source));
-                context.Types.DestinationType.ShouldBe(typeof(Dest));
-                context.ObjectMapper.ShouldBeNull();
-                context.MemberMap.ShouldBeNull();
-            }
-            else
-            {
-                context.MemberMap.SourceMember.Name.ShouldBe("Values");
-                context.MemberMap.DestinationName.ShouldBe("Values");
-                if(context.Types.Equals(new TypePair(typeof(int), typeof(int))))
-                {
-                    _calledForInt = true;
-                    context.ObjectMapper.ShouldBeOfType<AssignableMapper>();
-                }
-                else
-                {
-                    _calledForValues = true;
-                    context.ObjectMapper.ShouldBeOfType<ArrayCopyMapper>();
-                    context.Types.SourceType.ShouldBe(typeof(int[]));
-                    context.Types.DestinationType.ShouldBe(typeof(int[]));
-                }
-            }
-        }
     }
 
     public class When_using_a_type_converter : AutoMapperSpecBase
